[ARM] msm-nand: Add support for 4K Page Devices.
Add support for 4K page devices along with 2K page devices.
Since both the types of devices are to be supported, the
partition information that is obtained from SMEM or ATAGs
which contian starting block number and the number of blocks,
is now insufficient to determine the partition start and end
addresses. This computation is now deferred till the Nand
device type is identified during the Nand probe phase.

Signed-off-by: Murali Palnati <palnatim@qualcomm.com>