nohz: Fix ondemand governor not always ramping up on 100% load.
On systems that perform the context switch with interrupts enabled
(e.g. ARM) get_cpu_idle_time_us could get stuck returning the last
idle time when the cpu is busy. If an interrupt occurs while
switching to the idle thread, and this interrupt wakes up a thread,
need-resched will be set on the idle thread, but since we have not
finished switching to the idle thread tick_nohz_stop_sched_tick will
be called. (idle_cpu is true since the head of the runqueue is the
idle thread, but need_resched returns false since we are still
running in the context of the old thread)

Fix this by not calling tick_nohz_start_idle when returning from an
interrupt while the idle thread is in its schedule phase.
(!ts->inidle)

Change-Id: Ieb85db300ee91467ff3372503eeeb7c58b275c8d
Signed-off-by: Arve Hjønnevåg <arve@android.com>
Signed-off-by: Larry Bassel <lbassel@codeaurora.org>