[ARM] msm: Save test results across reads and writes
This module contains many tests that are launched by writing the test
name into the device node. Test results are retrieved by reading the
device node.

The test return value was being allocated in the open and freed in the
close. This meant that the typical open, read, close and open, write,
close of echo and cat deallocated the return value. This fix statically
allocates the return value so that its value is saved across echo and cat.

Signed-off-by: Zach Pfeffer <zpfeffer@quicinc.com>