pmem: add api for PMEM alloc and free from kernel space
int32_t pmem_kalloc(const size_t size, const uint32_t flags);

Attempts to allocate contiguous physical memory with the
specified alignment from the requested memory type.

This returns a physical address aligned to either 4k page boundary
or 1M boundary (if specified in the flags argument) or a negative
error number (i.e. TRUE when IS_ERR() is invoked on the return
value) if the allocation failed.

Size of the returned memory will be rounded up to the next full 4K
page.

Possible errors are -EINVAL (bad flag or size) and -ENOMEM (out of
memory).

Parameters:
size - is specified in bytes.

flags - are an or'ed bitmap of MEMTYPE and ALIGNMENT values which
define the type of memory requested and the alignment of the
allocated memory:

Valid values for MEMTYPE are:
PMEM_MEMTYPE_EBI1 (currently the only allowed type for in-kernel
allocations)

Valid values for ALIGNMENT are:
PMEM_ALIGNMENT_4K (the default) and PMEM_ALIGNMENT_1M

----------------------------

int32_t pmem_kfree(const int32_t physaddr);

This frees memory previously allocated by pmem_kalloc.
The entire size allocated starting at this address will be freed.

This returns 0 if the free succeeded or a negative kernel error
number if it failed.

Possible error values may be -EINVAL if a bad physaddr was provided.

"physaddr" must be a valid value that has been returned from
pmem_kalloc. This will be checked.

If, for some reason, a passed-in value manages to pass all the
internal checks but is an invalid value, and is submitted to the
freeing functions, undefined behavior may ensue, possibly including
corruption of memory arenas.

Example setup of the board files have been provided as part of this
submission.

Signed-off-by: Stephen Biggs <sbiggs@quicinc.com>