[ARM] msm: audio: 8k: Pass audio device ID to ACDB instead of HW path.
Update CAD driver to pass device ID to the ARM9 and query interface.
This change is required for the new Q6 firmware image
Audio ID is passed instead of HW path as in previous Q6 release.
This change is not backward compatible.
Side effect is noise will be heard in the output.

Signed-off-by: Ben Romberger <bromberg@quicinc.com>