USB: android gadget: adb function enable and disable support
performs adb enable or adb disable from the existing USB composition.

On adb enable request, if adb is not enabled in the existing
composition , this patch adds adb function to existing function map and
generates the new PID. If the new PID is supported list,it
performs comp switch.

On adb disable request, if adb is enabled in the existing composition ,
this patch removes the adb function from existing function map and
generates the new PID.
If the PID is in the supported list, it performs comp switch.

Signed-off-by: Anji Jonnala <anjir@qualcomm.com>