usb: msm_function: Reorganize suspend locks
HSUSB driver uses wakelocks and pm_qos to prevent platform
suspend. Calls to these locks(to unlock or lock) are scattered
all over the driver making it difficult to add any new rules.
For example; Code for requesting AXI clock frequency needs to
be added right after releasing these locks and it requires
changing at multiple places in the driver. In order to avoid
multiple changes and to organize the code better, defining
function which acquires/releases suspend locks. If there is
any new rule to be added it would be added to these functions
and rest of the driver would be untouched.

Signed-off-by: Vamsi Krishna <vskrishn@qualcomm.com>