video: Fix failure to fill in file pointer upon bad pmem return.
If the call to get_pmem_file in get_img function fails for any
reason, as it may well, the continuation of the function
to get the image file information was not filling in the file
pointer.  The resultant null file pointer was then being passed
to flush_pmem_file, another PMEM function.

Due to testing of new features in PMEM, kernel panics
were being caused by a temporary failure of PMEM to protect itself
adequately against other kernel drivers invoking its functions such
as flush_pmem_file that may have passed in a null file pointer as one
of its arguments.

Signed-off-by: Stephen Biggs <sbiggs@quicinc.com>