[ARM] msm: audio: 7k: Add a seperate ioctl for MBADRC
Background:
The earlier versions of DSP firmware supports only ADRC. The original
driver code from Google/HTC exposes this feature to user space throgh
the ioctl AUDIO_SET_ADRC. Now the latest DSP firmware supports MBADRC
feature.  Single band MBADRC is equivalent to ADRC.  DSP still retains
the ADRC command for backward compatibility and internally it converts
ADRC parameters to MBADRC.

Description:
Previous changes to this driver added support for MBADRC but exposed
this feature to user space through old ioctl AUDIO_SET_ADRC which is not
correct. So now adding a seperate ioctl AUDIO_SET_MBADRC keeping the
origial ADRC interface as is. With this driver can now accept ioctl
commands for both ADRC and MBADRC but internally driver issues only
MBADRC command to DSP. In case of ADRC it converts the parameters to
MBADRC before issuing DSP command. The reason for doing this conversion
inside driver is that in future even if DSP deprecates the support for
ADRC there will not be changes required in driver.

Signed-off-by: Suresh Vankadara <svankada@qualcomm.com>