EHCI: msm: Add/remove hcd, when mini A cable is connected/disconnected
When mini A cable is connected to the FSUSB port, VBUS is on
by the PMIC OTG driver. if hcd is already added by this time,
the ehci stack software dumps a lot of mesages on to the console.

To eliminate, these messages, add the hcd only, when mini A
cable is connected and remove the hcd, when mini A cable is
disconnected.

Signed-off-by: Velempati Chiranjeevi <c_cvelem@qualcomm.com>