EHCI: msm: suspend phy, when no device is connected on root hub
Current code doesn't suspend the phy, when a device is disconnected
from the root hub, rather informs the PMIC OTG driver that a device
has been disconnected.

With this fix, on cable disconnection from the root hub, gets RPC
call and as a part of that suspending the phy.

With the new modem changes to the PMIC OTG state machine, there is
no need to notify the remote device disconnection events from the
ARM11 side.

Signed-off-by: Velempati Chiranjeevi <c_cvelem@qualcomm.com>