[for -mm] mem_notify v6: memory_pressure_notify() caller
the notification point to happen whenever the VM moves an
anonymous page to the inactive list - this is a pretty good indication
that there are unused anonymous pages present which will be very likely
swapped out soon.

and, It is judged out of trouble at the following situations.
	o memory pressure decrease and stop moves an anonymous page to
	the inactive list.
	o free pages increase than (pages_high+lowmem_reserve)*2.

	ChangeLog:
		v5: add out of trouble notify to exit of
balance_pgdat().

Signed-off-by: Marcelo Tosatti <marcelo@kvack.org>
Signed-off-by: KOSAKI Motohiro <kosaki.motohiro@jp.fujitsu.com>