pmem: add test module for kernel api
The init portion will return an error and not stay loaded if there
is any problem with the test.  If there is no problem, an OK status
is returned via the insmod or modprobe call and the module stays
loaded in the kernel until it is removed via rmmod or modprobe.

This set of tests is modeled after the userspace _pmem_test.c set of
stress tests. This module does two nominal tests, testing first an
allocation of 1 Megabyte at 4k page alignment, write and read-back
and free, and the same cycle again at 1 Megabyte alignment and then
the full nominal test done for 2 Megabytes at both alignments.

A set of error parameters are passed in as expected failures.

There is also an adversarial test where there are two allocations
at once, one of 1 Megabyte and the other after at 64 K with the same
read/write and free tests, for both types of alignment.

Signed-off-by: Stephen Biggs <sbiggs@quicinc.com>