[ARM] msm: audio: 7k: Allow reading residual buffers in pcm_in driver
In current implementation user is not allowed to call read once
AUDIO_STOP is called to stop recording audio. But around one second of
data can be missed in a case where some recorded data is left out with
the driver. To avoid this the read call is allowed even after stop
untill the recorded buffer count becomes zero.

Signed-off-by: Sriranjan Srikantam <c_ssrika@qualcomm.com>