spi_qsd: Add support for loopback mode
Support for loopback mode makes it easier to test the SPI driver and
controller hardware without requiring that specific slave devices are
connected to the board under test.  Loopback mode allows us to be
slave independent.  In addition, loopback mode can be valuable when
using DMA for data transfers.  It is possible to exercise large data
transfers that might not other be supported by the slave devices on hand
because these devices may have limitations on the amount of data that
can be sent or received.

Signed-off-by: Kenneth Heitke <kheitke@quicinc.com>