usb: msm_otg: Add platform suspend locks support
Wakelock and PMQOS parameters are used to prevent platfrom suspend
when usb is active. Both wakelocks and PMQOS are needed because, 8K
platform does not honour wakelocks.
Platform suspend locks(wakelocks and pm qos) will not be released
at USB bus-suspend. If platform suspend is allowed at usb bus suspend
it cannot not meet chrip timing requirement at bus reset. But for
power test automation, which is generally runs through usb. It would
never go to platform suspend as USB is always active. For power test
automation it is assumed that, testcase will set a flag for wakelocks
release and switch to a composition which allows bus-suspend. USB otg
will check for this flag at the end of low power mode function and
and will release wakelocks if this flag is set.
Add Debugsfs entry to control platform suspend locks release
at low power mode function.

Signed-off-by: Vamsi Krishna <vskrishn@qualcomm.com>