usb: msm_gadget: Add Charger support
Peripheral device is allowed to draw pre-specified amount of power
from bus at different stages. When device is attached it is allowed
to draw 100mA of current, when bus is suspended it is allowed to
draw 2.5mA and when device is configured it can draw upto the value
specified in configuration descriptor. When wall charger is attached
it can draw 1500mA of current.
Add wall charger detection. Wall charger is detected if Dp/Dn pins
are shorted. Add calls to the otg at different device states to
communicate how much power can be pulled from bus.
Driver also provides sysfs entries for charger type and charger
current. Charger type can be one of the below:
0 - HOST PC is connected
1 - Wall Charger is connected
2 - Nothing is connected
charger current specifies the mA requested to be drawn from usb bus.

Signed-off-by: Vamsi Krishna <vskrishn@qualcomm.com>