logger: Modify priority checking
Switch order of priority checking so that per-tag is enabled
regardless of the global enable for a particular logger.  That way,
a particular tag can continue to log while the rest are filtered out.

As part of this change, when the global priority per logger type,
all tags currently attached to that logger also will have their
priority reset to the global logger priority.

All new tags now take on the current global priority of the logger
type they are to be attached to.

Perform minor cleanups to the code.

CRs-fixed: 182025

Signed-off-by: Stephen Biggs <sbiggs@quicinc.com>