Input: gpio_event: Allow multiple input devices per gpio_event device
This is needed to support devices that put non-keyboard buttons in
the keyboard matrix. For instance several devices put the trackball
button in the keyboard matrix. In this case BTN_MOUSE should be
reported from the same input device as REL_X/Y.

It is also useful for devices that have multiple logical keyboard in
the same matrix. The HTC dream has a menu key on the external keyboard
and another menu key on the slide-out keyboard. With a single input
device only one of these menu keys can be mapped to KEY_MENU.

Change-Id: Ib91f6c5b642d96555c273c33b08efa979b6aabc1
Signed-off-by: Arve Hjønnevåg <arve@android.com>
[lbassel@codeaurora.org: fix minor merge conflicts]
Signed-off-by: Larry Bassel <lbassel@codeaurora.org>