ipv4: fix kernel crash when data is shut down
The tcp_v4_nuke_addr holds a spinlock and ends up calling
inet_unhash that tries to lock the same spinlock leading to
deadlock. The fix releases the spinlock just before calling
tcp_done and retraverses the list from the beginning.

This does increase the complexity to n^2 since it traverses the entire
list for each applicable socket.This could potentially be a problem if
there a lot of sockets belonging to different interfaces. On mobile
platform this wont be an issue since it has a single interface.

CRs-fixed: 187180
Signed-off-by: Abhijeet Dharmapurikar <adharmap@quicinc.com>