msm: usb: Add support for charger version 2.1
7630 target uses charger version 2.1.  This change make
it easier to add server versions for compatibility and
ensures that we verify compatibility with both the local
code and the remote server version.

The changes also make the usb driver with subsequent
minor revision changes such as 1.x or 2.x and makes
it easier to add subsequent major number support.

Signed-off-by: Richard Ruigrok <rruigrok@quicinc.com>