msm: adsp: Decrease polling interval in cmd/msg processing functions
In the function msm_adsp_write, ADSP driver waits for maximum of
100usecs with 10usecs interval until ADSP is ready to accept a
command and waits for maximum of 5msecs with 1msecs interval until
ADSP responds to interrupt. Similarly the driver waits for maximum
of 100usecs with 10usecs interval until ADSP is ready for message
processing inside the function adsp_get_event which gets called from
irq handler.

In the above three cases the duration of the delay, all of which are
busy wait loops, before  polling introduces unnecessary delay while
writing a command and message processing.  Hence reduce the polling
interval to 2 usecs and adjust the loop counts accordingly to improve
performance.

Signed-off-by: Suresh Vankadara <svankada@qualcomm.com>