msm: Fix bugs in oem_rapi driver
Size of each data item transmitted over RPC should be in multiple
of 4 bytes. Strings passed are not padded to make their size a
multple of 4. Pad the strings with 0s to make its size a multiple of 4.

As per the marshalling protocol, the size of variable sized array should
only be sent if it is a valid pointer. Correct code to send output_size
only if output_valid is true.

Buffer pointer is not incremented after reading out_len.
Correct buffer increment after reading out_len parameter.

Signed-off-by: Niranjana Vishwanathapura <nvishwan@quicinc.com>