msm: remote-spinlock: Use SWP-based locking with ARM11 apps cores
Use a remote spinlock mechanism based on the SWP instruction instead of
one based on the local spinlock's LDREX/STREX implementation.  This is
required for MSM7XXX SoC's equipped with an ARM11 application processor.
The memory controllers in these SoC's do not support tagging shared
memory addresses for exclusive access and therefore LDREX and STREX cannot
be used for locking between the modem and apps processors.

For SoC's supporting exclusive tagging of shared memory, as well as for
spinlocks local to the application processor, the LDREX/STREX
implementation is preferable and therefore still used.

Signed-off-by: Matt Wagantall <mattw@quicinc.com>