mfd: pmic8058: add correct protection and other fixes
The functions in irq_chip are called with local interrupts disabled and
generic_handle_irq also requires local interrupts disabled. Thus, add
local_irq_disable in the interrupt thread to prevent race condition for
the private data and to meet the above requirement. This change is ok
for calling i2c_transfer because it is able to handle the context with
interrupts disabled.

Also added were bug fixes and gpio/keypad parameter fixes.

Signed-off-by: Willie Ruan <wruan@quicinc.com>