pmem: fix spinlock recursion messages
Fix worrisome (but ultimately benign, at least in this case) output
from kernel related to spinlock debugging as a result of attempting
to acquire more than one spinlock of the same type as a result of
PMEM connect action.

Two different file pointers have different internal data
pointers and different spinlock instances. But, the spinlocks are
of the same class, hashed based on their initial name which is the
same for all internal file data spinlocks.  Thus, grabbing more than
one internal data spinlock causes kernel debugging code to activate
due to restrictions in the kernel deadlock detection debug code on
acquisition of more than one spinlock of the same type or class.

Fix here is to save needed data from the private data of the source
file into a local variable under spinlock, then relinquish that
spinlock immediately before then acquiring the spinlock of the other
file pointer to store that saved data in the private data of the
other file pointer.

Signed-off-by: Stephen Biggs <sbiggs@quicinc.com>