msm_fb: mdp4: Add support for 128 MHz AXI bus
Switching the AXI bus from 64 Mhz to 128 MHz causes the MDP to
hang during large data transfers on the msm7x30 SURF.

Fix:

-Reduce the DMA_P burst size from 16 to 8.
-Reduce the number of pending commands for the AXI port to 3.

The data shows that either of these changes alone
allows the display to work correctly, but for the time being we will set
both to be conservative.

According to the hardware folks, if the MDP clock is less than
the AXI clock, then we must use a max of three pending requests
-- otherwise we should be able to use 8.  We should make a change
in the future to support this with proper testing.

Signed-off-by: Michael Bohan <mbohan@quicinc.com>