pmic8058-keypad: add workaround to fix last row problem
There is a hw bug in PM8058 to generate multiple interrupts when a key
on last row is pressed. The workaround is to configure max rows, but
only process actual rows of keys.

Also changed are the code to configure column and row bits based on
resource numbers and to use row hold time from resource.

Signed-off-by: Willie Ruan <wruan@quicinc.com>