usb: msm_function: Changing the attributes of the autoresume file.
We have to write '1' into the file autoresume of sysfs, for the remote
wakeup feature of USB to work. Implemented it as a UI option in Qualcomm
settings app as required for the USB compliance to pass. We have to
change the attributes of the autoresume file, so that Qualcomm settings
app can write into it.

CRs-fixed: 176588
Signed-off-by: Mahesh Vojjala <c_mvojja@quicinc.com>