msm: audio: 8k: Converted device ID before device volume caching
When the device volume caching code was modified to index
into the caching structure using the internal device ID
the CAD device ID was not converted into the internal
device ID before getting the dB range of the device.
Now the proper dB range will be used for each device.

CRs-fixed: 193175
Signed-off-by: Ben Romberger <bromberg@quicinc.com>