msm: Add mapping of oncrpc program numbers to program names.
Symbols represent oncrpc programs that may be exported by the
remote processor.

smd_rpc_sym generated from perl script creates a list of
symbols. Undefined symbols are filled such that there are no gaps
and a simple index can be use to dereference the table to obtain
the program number.

By definition, program starting with 0x30 are forward APIS
and program starting with 0x31 are callbacks.
This table takes advantage of this fact to reduce the number of
symbols by half by prepending each symbol with "CB "
and skipping the first 3 characters for forward APIS names.

Signed-off-by:  Richard Ruigrok <rruigrok@quicinc.com>