mfd: marimba: Fix the value being written to i2c write
Fixed the wrong values being written to the Marimba registers.
The value the subdevice written to the register was wrong
because the write api expected the values being passed to
be masked first but this is not consistent with all the
sub devices using this api. Subdevices should only
pass the value it needs to write and the mask for the
appropriate bits in the register. The core driver api will
use this mask and set the registers bits.

Acked-by: Siddartha Mohanadoss <smohanad@qualcomm.com>
Signed-off-by: Michael Bohan <mbohan@quicinc.com>