spi_qsd: Add support for SPI PCLK
The SPI core has two clocks: one for the register interface and one for
the master shift register.  If no PCLK is provided to the driver, then
the two clocks are combined and treated as a single clock.  If the PCLK
is provided, then the two clocks are treated independently and each clock
can be turned on or off as needed.

Signed-off-by: Kenneth Heitke <kheitke@quicinc.com>