diag: Performance Improvement for DIAG char driver
Add data aggregation in the DIAG driver. Previously DIAG used to send
each data packet to USB individually. This resulted in low throughput.
With this patch, now the DIAG driver will aggregate data before shipping
it to USB. It will wait until it reaches a max number of bytes or a
timer goes off. On reaching either of threshold, it will call USB write
API. Also, it will write out as soon as a packet response comes in, as
they are time sensitive data. This has resulted in much improved
throughput.

Currently this aggregation is for Apps Processor traffic only. The modem
traffic is routed as is.

Signed-off-by: Shalabh Jain <shalabhj@quicinc.com>