mmc: msm_sdcc: Make polling enabled status as per host data.
When CONFIG_HAS_EARLYSUSPEND is defined, the polling enabled/disabled state
of the slot is currently tracked using a single global variable for all the
slots. This results in improper handling of polling when the system resumes
back from the suspend state.

This patch moves this global variable into msmsdcc_host structure, which is
a host specific data structure. With this since each slot gets its own copy
of polling status flag, corruption of polling status, during suspend/resume
cycles are prevented.

CRs-fixed: 194991

Signed-off-by: Murali Palnati <palnatim@qualcomm.com>