yaffs: Read performance optimization
YAFFS supports the idea of "chunk groups" which allow the tnode entries
to be smaller. For example, if we have 2^17 chunks but only 16 bits of
tnode, then each chunk group is 2 ^ (17-16) = 2 entries. If chunk groups
are being used then we don't know exactly which chunk is being
references and we have to search for the chunk by reading the tags. If
you're using wide tnodes then the chunk group is size 1 then there is no
need to search since we know which it is.  We thus don't need to do the
read-and-search.

On our phone target hardware, write performance remained unchanged, but
read performance improved by appx 35%, measured with LMDD with the
OnDemand CPU governor.

Signed-off-by: Charles Manning <cdhmanning@gmail.com>
Signed-off-by: David Neiss <dneiss@quicinc.com>