usb: msm_function: Remove pmic interrupt handler registration
Because of H/W bug in 7201a usb phy  cannot detect bus reset when
usb phy is in low power mode and TCXO is turned off. Inorder to
workaround this, ulpi_data0(D+ line) is routed to PMIC MPP 15
interrupt. Apps USB driver has to make an RPC call inorder to
enable and listen to this interrupt while going to low power mode
On the other hand; when usb bus suspended, driver should be able
to respond to bus reset within 3.5ms. When TCXO is off, estimated
driver latency in worst case is 5ms. Driver will not able to
respond to bus reset within specified duration if TCXO is off.
To prevent this, usb driver does not vote for power collapse
during bus supsend.
So USB driver not allowing TCXO shutdown during bus suspend and
there is no need of relying on pmic interrupt handler.

Signed-off-by: Vamsi Krishna <vskrishn@qualcomm.com>