msm: audio: 7k: Remove unnecessary parameter from AAC adec params structure
Remove the unnecessary parameter(dual_mono_mode) in AAC adec params
configuration structure which was added previously by mistake. This was
corrupting the channel configuration parameter as this extra parameter
is resulting in overflow of the last parameter in this structure.

This bug was found while testing the BSAC feature as DSP uses the
channel configuration parameter in case of BSAC and this parameter is
not read by DSP for other formats. Hence the bug was not captured
previously.

Signed-off-by: Sriranjan Srikantam <c_ssrika@qualcomm.com>